Rename the PowerOffice namespace to Invoicing
