+++ src/EventManagement.Services/Invoicing/PowerOfficeService.cs
@@ -12,7 +12,7 @@ using losol.EventManagement.Infrastructure;
 using Microsoft.Extensions.Options;
 using static losol.EventManagement.Domain.PaymentMethod;
 
-namespace losol.EventManagement.Services.PowerOffice {
+namespace losol.EventManagement.Services.Invoicing {
     public class PowerOfficeService : IInvoicingService {
 
         private readonly Go api;
