+++ src/EventManagement.Web/Startup.cs
@@ -19,7 +19,7 @@ using losol.EventManagement.Services.DbInitializers;
 using losol.EventManagement.Services.Messaging;
 using losol.EventManagement.Web.Services;
 using losol.EventManagement.Services.Messaging.Sms;
-using losol.EventManagement.Services.PowerOffice;
+using losol.EventManagement.Services.Invoicing;
 using losol.EventManagement.Config;
 using losol.EventManagement.Web.Config;
 using losol.EventManagement.Web.Extensions;
@@ -48,7 +48,7 @@ namespace losol.EventManagement
             });
 
 
-                // sqlite: options.UseSqlite(Configuration.GetConnectionString("DefaultConnection"))); 
+                // sqlite: options.UseSqlite(Configuration.GetConnectionString("DefaultConnection")));
 
             services.AddIdentity<ApplicationUser, IdentityRole>(config =>
             {
@@ -68,7 +68,7 @@ namespace losol.EventManagement
                 });
             }
             */
-            
+
 
             // Set password requirements
             services.Configure<IdentityOptions>(options =>
@@ -85,18 +85,18 @@ namespace losol.EventManagement
             CultureInfo.DefaultThreadCurrentCulture = cultureInfo;
             CultureInfo.DefaultThreadCurrentUICulture = cultureInfo;
 
-            
+
             services.AddAuthorization(options =>
             {
                 options.AddPolicy("AdministratorRole", policy => policy.RequireRole("Admin", "SuperAdmin"));
             });
-            
+
             services.AddMvc()
                 .AddRazorPagesOptions(options =>
                 {
                     options.Conventions.AuthorizeFolder("/Account/Manage");
                     options.Conventions.AuthorizePage("/Account/Logout");
-                    
+
                     options.Conventions.AuthorizeFolder("/Admin", "AdministratorRole");
                     options.Conventions.AddPageRoute("/Events/Details", "events/{id}/{slug?}");
 
@@ -173,7 +173,7 @@ namespace losol.EventManagement
             {
                 services.AddTransient<IInvoicingService, MockInvoicingService>();
             }
-            
+
 
 			// Register our application services
 			services.AddScoped<IEventInfoService, EventInfoService>();
@@ -222,7 +222,7 @@ namespace losol.EventManagement
             /*
             if (env.IsProduction()) {
                 var options = new RewriteOptions()
-                .AddRedirectToHttps(); 
+                .AddRedirectToHttps();
                 app.UseRewriter(options);
             }
              */
